feat: Break Stream into Mock and Gmock implementations
